[malonaz/ai] - add per-role message counts and per-model usage to chat, widen quantity to int64

+++ go/ai/ai_service/provider/deepgram/stream_text_to_speech.go
@@ -179,10 +179,10 @@ func (c *Client) StreamTextToSpeech(
 	modelUsage := &aipb.ModelUsage{
 		Model: configuration.Model,
 		InputCharacter: &aipb.ResourceConsumption{
-			Quantity: int32(billableCharacterCount),
+			Quantity: int64(billableCharacterCount),
 		},
 		OutputSecond: &aipb.ResourceConsumption{
-			Quantity: int32(totalDuration.Seconds()),
+			Quantity: int64(totalDuration.Seconds()),
 		},
 	}
 	if err := srv.Send(&aiservicepb.StreamTextToSpeechResponse{
